Refactor and extend gating strategy and change results report
* Change and extend gating strategy and fix draw trajectories

* Some code refactor & report yolo detections instead of kf means

+++ strong_sort/sort/detection.py
@@ -27,7 +27,7 @@ class Detection(object):
     """
 
     def __init__(self, class_id, tlwh, confidence, feature):
-        self.tlwh = np.asarray(tlwh, dtype=np.float)
+        self.tlwh = np.asarray(tlwh, dtype=np.float32)
         self.confidence = float(confidence)
         self.class_id = int(class_id)
         self.feature = np.asarray(feature.cpu(), dtype=np.float32)
